+++ src/components/kakaoLoginButton/KakaoLoginButton.tsx
@@ -1,43 +1,15 @@
-import React, { useEffect } from 'react';
-import { signIn, signOut, useSession } from 'next-auth/react';
+import React from 'react';
+import { signIn } from 'next-auth/react';
 import { css } from '@emotion/react';
 
-import { LOCAL_STORAGE_KEY } from '~/constants/storage';
-import { post } from '~/libs/api';
+import useKakaoLogin from '~/components/kakaoLoginButton/useKakaoLogin';
 import colors from '~/styles/color';
 import { BODY_1 } from '~/styles/typo';
 
-interface Token {
-  token_type: string;
-  access_token: string;
-}
-
 const KakaoLoginButton = () => {
-  const { data: session } = useSession();
+  const { logOutHandler, isLoginState } = useKakaoLogin();
 
-  const getTokenHandler = async () => {
-    const token: Token = await post('/oauth/kakao', {
-      nickname: session?.user?.name,
-      email: session?.user?.email,
-    });
-
-    localStorage.setItem(LOCAL_STORAGE_KEY.accessToken, token.access_token);
-  };
-
-  const logOutHandler = () => {
-    signOut();
-    localStorage.removeItem(LOCAL_STORAGE_KEY.accessToken);
-  };
-
-  useEffect(() => {
-    if (session) {
-      getTokenHandler();
-    }
-
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [session]);
-
-  if (session) {
+  if (isLoginState) {
     return (
       <div css={KakaoLoginWrapper}>
         <button type="button" onClick={logOutHandler}>
